feat: full-stack discussions (create)
